refactor: Update authentication dependencies and enhance test structure
* Refactored the `get_current_user` function to use a more explicit credentials dependency.
* Improved test setup by adding a database cleanup fixture to ensure a clean state before each test.
* Updated tests to utilize dependency injection for the TestClient, enhancing test readability and maintainability.
* Moved utility functions for user creation and authorization headers to a new `utils.py` file for better organization.
